Refactor package structure into modular directories
