+++ socialtools_app/app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <!-- Required for starting foreground services on Android 9+ -->
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
     <application
-        android:name=".App"
+        android:name=".core.App"
         android:allowBackup="true"
         android:usesCleartextTraffic="true"
         android:icon="@mipmap/ic_launcher_foreground"
@@ -20,14 +20,14 @@
                 android:name="android.support.FILE_PROVIDER_PATHS"
                 android:resource="@xml/provider_paths" />
         </provider>
-        <activity android:name=".MainActivity" android:exported="true">
+        <activity android:name=".ui.MainActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <service
-            android:name=".PostService"
+            android:name=".core.services.PostService"
             android:exported="false" />
     </application>
 </manifest>
