+++ socialtools_app/app/src/main/java/com/cicero/socialtools/core/App.kt
@@ -1,4 +1,4 @@
-package com.cicero.socialtools
+package com.cicero.socialtools.core
 
 import android.app.Application
